Move `DomainStatus` to shared module
