+++ cli/src/api/domain/remove.rs
@@ -1,4 +1,4 @@
-use crate::api::domain::add::DomainStatus;
+use crate::api::domain::DomainStatus;
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct Request {
@@ -10,6 +10,7 @@ pub struct Request {
 pub struct Response {
     /// Domain name (e.g. example.com)
     pub domain_name: String,
+    /// Domain status (e.g. Pending)
     pub status: DomainStatus,
     /// Whether a deploy is needed to complete removal
     pub requires_deploy: bool,
