+++ cli/src/api/domain/status.rs
@@ -1,4 +1,4 @@
-use crate::api::domain::add::DomainStatus;
+use crate::api::domain::DomainStatus;
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct Request {
@@ -10,6 +10,8 @@ pub struct Request {
 pub struct Response {
     /// Domain name (e.g. example.com)
     pub domain_name: String,
+    /// Domain status (e.g. Pending)
     pub status: DomainStatus,
+    /// Last time the domain was checked
     pub last_checked_at: Option<chrono::DateTime<chrono::Utc>>,
 }
